More overall small changes. Fixed Surface problem.
Fixed the problem where I had to force the Surface to become public from
the titlescreen test. Maybe I didn't know how to use the features but it
seems that a class can access a private member of it's own instance,
didn't know that so I kept trying tricks and going around it with
pointers and friend xD

Also a small style change from using bracketed #include's for libraries
to using angle brackets.