FontSurface Blended mode + fixed EventHandler.
Added support for rendering text in Blended mode for the FontSurface
class. Also solved the issue about the Color being statically or
dynamically allocated. It's a small dirty trick you could say, but does
the job for now ^^

Also fixed the EventHandler class in relation to the problems it gave me
when testing titlescreen. Switched from using pointers to SDL_Event to
SDL_Event structs and dereferencing them internally for the Event
functions.