Tons of progress on the movingaround test.
Got around to implementing the game states as a base class and coded the
menu, exit and most of the settings state. also figured out how to use
my singleton :D Of course, everything is really badly written and stuff
but I think the test will help much.

Also, added the virtual OnKeyUp() function to the event handler seeing
as the movingaround test needs it. Also, the FontSurface class now
supports changing/setting of color on them.